kpatch-build: cross-compile: determine ARCH from $VMLINUX
Since an existing vmlinux is required for successful kpatch-build,
use it to determine $ARCH, instead of assuming target is $(uname -m).

For cross-platform builds, ARCH must be passed to the underlying
make, but that is _not_ the local $ARCH of kpatch-build.
It can be passed thru with --make-opt=ARCH=arm64 for example,
whereas the machine arch name elf kpatch use is aarch64

Signed-off-by: Pete Swain <swine@google.com>