DO_NOT_UPSTREAM: kpatch-cc should ignore usr/*
Ignore all changes to usr/* which can never be patched

The failure trigger here was gbuild-9.3's with
commit 7d2a80a5ccb1 ("gbuild: kinit: Support specifying root by NGUID")
of gbuild/klibc/usr/kinit/find-disk/... which builds in subdir,
with a target object path usr/kinit/find-disk/find-disk.o

Whether built from top dir, or subdir, the prefix usr/ is
always a sign that kpatch-cc should not process the object.

This will have the effect of failing to report as non-patchable
any attempt to change (eg) usr/initramfs_data.S but I don't
don't regard that as an issue - such a livepatch request would
no longer fail in the tool, but would never pass human review.

Upstreamable: n/a for _this_ issue, but adding a site-specific
    exclusion pattern to kpatch-cc to avoid forking would
    be a good follow-up, and that's upstreamable.
    We'd export it from generate-klp, to steer kpatch-cc

Tested: now builds working klp modules with 930.6
    which includes gbuild-9.3